Bizlogic implementation

+++ domain/src/main/java/com/nerevar/domain/Note.kt
@@ -1,3 +1,3 @@
 package com.nerevar.domain
 
-data class Note(val id: Int, val title: String)
+data class Note(val id: Long, val title: String)
